examples/memory-backend-agent/pyproject.toml
[project]
name = "memory-backend-agent"
version = "0.1.0"
description = "Deep Agent example using MemoryBackend for path-keyed storage (PathMemoryStore)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.5",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.6",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.urls]
Repository = "https://github.com/langchain-ai/deepagents/tree/main/examples/memory-backend-agent"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
